feat(FN-995): catch errors and use central api to get bank id
