--- dtfs-central-api/src/v1/controllers/portal/utilisation-report-service/utilisation-report-upload.controller.js
@@ -1,16 +1,0 @@
-const db = require('../../../../drivers/db-client');
-
-const getPaymentOfficerTeamDetails = async (req, res) => {
-  const { bankId } = req.params;
-  try {
-    const bankCollection = await db.getCollection('banks');
-    const filteredBank = await bankCollection.findOne({ id: bankId });
-    res.status(200).send(filteredBank);
-  } catch (error) {
-    console.error('Unable to get utilisation reports %s', error);
-  }
-};
-
-module.exports = {
-  getPaymentOfficerTeamDetails,
-};
