+++ portal/templates/utilisation-report-service/utilisation-report-upload/confirmation.njk
@@ -7,7 +7,7 @@
             <div
                 class="govuk-panel govuk-panel--confirmation">
                 <h1 class="govuk-panel__title" data-cy="confirmation-main-heading">
-                    {{ reportMonthYear }} GEF report sent to UKEF
+                    {{ reportPeriod }} GEF report sent to UKEF
                 </h1>
             </div>
             <p class="govuk-body" data-cy="paragraph">A confirmation email has been sent to {{ bankEmail }}.</p>
